do-caveats.sh: new tools script (#249)
Does the caveats of all formula which are currently installed. This
script is somewhat slow because it calls brew every for every formula to
see if it is installed. If its necessary to make this faster brew can be
called once with all the formula name, but it would be more difficult to
see which formula are installed and what their caveats are.